fix(ValueFunction): Proper handling of non interger keys
Keys will be not only integer but can be for examples the
Grid point index as an integer array. That case was not
handled by the equals and isClose functions properly as
they were only checking regular integer array indexes,
meaning that the actual set values of the value function
where never checked.